+++ CHANGELOG.md
@@ -19,7 +19,8 @@
 * issue#171: Better TOP/Bottom5 panel styles
 * issue#172: Fix PHP notice when favourite graph exists and timespan changes
 * feature#161: Add busiest panels (DS Stats)
-* feature#164: Line graph add zoom, pie graph add tooltip percentage 
+* feature#164: Line graph add zoom, pie graph add tooltip percentage
+* feature#182: Normal or wide (more panels on a line) display
 
 --- 3.0.3 ---